Add basic support for monitor mirroring, fix --select, etc.
Basic support for monitor mirroring is added. The implementation I have is okay, but it can look weird if the monitor is smaller than the terminal.
Key input is now alterable in the config file.
The --select option is fixed. Before, it would still select the specified space, but the viewport wouldn't update for some reason.